explore, explore admin loop

+++ pages/exploreadmin.php
@@ -1,3 +1,5 @@
+<!DOCTYPE html>
+
 <head>
     <title>Tubirit | Explore</title>
 
@@ -6,6 +8,7 @@
     <link rel="stylesheet" href="https://unpkg.com/boxicons@latest/css/boxicons.min.css">
     <link rel="stylesheet" href="/css/style.css">
 </head>
+
 <body>
     <div class="container-fluid py-5">
         <div class="container text-start">
@@ -33,45 +36,50 @@
 
         <!--RECOMMENDATION LIST-->
         <div class="container">
-                    <?php
-                        include('./inc.koneksi.php');
-                        require_once('./class/class.buku.php');
+            <div class="row">
+                <?php
+                include('./inc.koneksi.php');
+                require_once('./class/class.buku.php');
 
-                        $objBuku = new Buku();
-                        $arrayResult = $objBuku->SelectAllBuku();
-                        //$arrayResult = $objBuku->SelectAllBukuByGenre();
+                $objBuku = new Buku();
+                $arrayResult = $objBuku->SelectAllBuku();
 
-                        if (count($arrayResult) == 0) {
-                            echo '<p>Tidak ada data!</p>';
-                        } /*else if(isset($_GET['selectgenre'])){	
-                            $objBuku->selectgenre = $_GET['selectgenre'];	
-                            $objBuku->SelectAllBukuByGenre();
-                        
-                        } */else {
-                            $no = 1;
-                            foreach ($arrayResult as $dataBuku) {
-                                echo '<div class="row">';
-                                echo '<div class="col-3 card mb-3 mx-2">';
-                                echo '<div class="row g-0">';
-                                echo '<div class="col-sm-6 py-2">';
-                                echo '<img class="img-fluid rounded" alt="..." src="./img/books/' . $dataBuku->cover . '"></div>';
-                                echo '<div class="col-sm-6">';
-                                echo '<div class="card-body">';
-                                echo '<h5 class="card-title">' . $dataBuku->judul . '</h5>';
-                                echo '<h6 class="card-title">' . $dataBuku->penulis . '</h6><br>';
-                                echo '<p>' . $dataBuku->namagenre . '<br>'. $dataBuku->penerbit . '<br>'. $dataBuku->halaman . ' halaman<br>' . $dataBuku->tahun . '</p></div></div>';
-                                echo '<p class="card-text">' . $dataBuku->summary . '</p>';
-                                echo '<div class="card-body">
-                                        <div class="float-right btn-group btn-group-sm">
-                                            <a href="dashboardadmin.php?p=buku&idbuku='.$dataBuku->idbuku.'" name="btnSubmit" class="btn btn-primary tooltips" data-placement="top" data-toggle="tooltip" data-original-title="Edit"><i class="bx bx-pencil"></i></a>
-                                            <a href="dashboardadmin.php?p=deletebuku&idbuku='.$dataBuku->idbuku.'" class="btn btn-secondary tooltips" 
-                                            onclick="return confirm(\'Apakah anda yakin ingin menghapus?\')" data-placement="top" data-toggle="tooltip" data-original-title="Delete"><i class="bx bx-trash"></i></a>
-                                        </div>
-                                    </div></div></div></div>';
-                                $no++;
-                            }
-                            } 
-                    ?>           
+                if (count($arrayResult) == 0) {
+                    echo '<p>Tidak ada data!</p>';
+                } else {
+                    $no = 1;
+                    foreach ($arrayResult as $dataBuku) {
+                ?>
+                        <div class="col card mb-3 mx-2">
+                            <div class="row g-0">
+                                <div class="col-sm-6 py-2">
+                                    <img class="img-fluid rounded" alt="<?php echo $dataBuku->judul; ?>" src="./img/books/<?php echo $dataBuku->cover; ?>">
+                                </div>
+                                <div class="col-sm-6">
+                                    <div class="card-body">
+                                        <h5 class="card-title"><?php echo $dataBuku->judul; ?></h5>
+                                        <h6 class="card-title"><?php echo $dataBuku->penulis; ?></h6>
+                                        <br>
+                                        <p><?php echo $dataBuku->namagenre; ?><br><?php echo $dataBuku->penerbit; ?><br><?php echo $dataBuku->halaman; ?> halaman<br><?php echo $dataBuku->tahun; ?></p>
+                                    </div>
+                                </div>
+                                <p class="card-text"><?php echo $dataBuku->summary; ?></p>
+                                <div class="card-body">
+                                    <div class="float-right btn-group btn-group-sm">
+                                        <a href="dashboardadmin.php?p=buku&idbuku=<?php echo $dataBuku->idbuku; ?>" name="btnSubmit" class="btn btn-primary tooltips" data-placement="top" data-toggle="tooltip" data-original-title="Edit"><i class="bx bx-pencil"></i></a>
+                                        <a href="dashboardadmin.php?p=deletebuku&idbuku=<?php echo $dataBuku->idbuku; ?>" class="btn btn-secondary tooltips" onclick="return confirm(\'Apakah anda yakin ingin menghapus?')" data-placement="top" data-toggle="tooltip" data-original-title="Delete"><i class="bx bx-trash"></i></a>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                <?php
+                        $no++;
+                    }
+                }
+                ?>
+            </div>
         </div>
     </div>
-<body>
+</body>
+
+</html>
